[Brightbox] Extract Compute::Shared to own file
Seeing a lot of churn in lib/fog/brightbox/compute.rb because of adding
new requests or models as well as any implementation changes.

This moves the common behaviour between Compute's Real and Mock
classes into their own file to cut down on the noise when refactoring or
merging.
